requirements_cpu.txt
torch==2.3.1+cpu
torchvision==0.18.1+cpu
pillow~=11.0.0
flask~=3.1.0
matplotlib>=3.7.2
numpy
tqdm